backend/requirements.txt
# FastAPI
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
pyodbc>=5.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Azure SDKs
azure-storage-blob>=12.19.0
azure-ai-vision-imageanalysis>=1.0.0

# AI Agents
crewai[azure-ai-inference]>=1.7.0

# Excel
openpyxl>=3.1.0

# Image processing
Pillow>=10.0.0

# HTTP
httpx>=0.27.0

# Config
python-dotenv>=1.0.0

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.4.0
